+++ ICD.Connect.Conferencing/Proxies/AbstractProxyDialingDeviceControl.cs
@@ -1,0 +1,171 @@
+﻿using System;
+using System.Collections.Generic;
+using ICD.Common.Properties;
+using ICD.Common.Utils.EventArguments;
+using ICD.Common.Utils.Extensions;
+using ICD.Connect.Conferencing.ConferenceSources;
+using ICD.Connect.Conferencing.EventArguments;
+using ICD.Connect.Devices.Proxies.Controls;
+using ICD.Connect.Devices.Proxies.Devices;
+
+namespace ICD.Connect.Conferencing.Proxies
+{
+	public abstract class AbstractProxyDialingDeviceControl : AbstractProxyDeviceControl, IProxyDialingDeviceControl
+	{
+		public event EventHandler<ConferenceSourceEventArgs> OnSourceAdded;
+		public event EventHandler<BoolEventArgs> OnDoNotDisturbChanged;
+		public event EventHandler<BoolEventArgs> OnAutoAnswerChanged;
+		public event EventHandler<BoolEventArgs> OnPrivacyMuteChanged;
+
+		private bool m_AutoAnswer;
+		private bool m_PrivacyMuted;
+		private bool m_DoNotDisturb;
+
+		#region Properties
+
+		/// <summary>
+		/// Gets the AutoAnswer state.
+		/// </summary>
+		public bool AutoAnswer
+		{
+			get { return m_AutoAnswer; }
+			[UsedImplicitly]
+			private set
+			{
+				if (value == m_AutoAnswer)
+					return;
+
+				m_AutoAnswer = value;
+
+				OnAutoAnswerChanged.Raise(this, new BoolEventArgs(m_AutoAnswer));
+			}
+		}
+
+		/// <summary>
+		/// Gets the current microphone mute state.
+		/// </summary>
+		public bool PrivacyMuted
+		{
+			get { return m_PrivacyMuted; }
+			[UsedImplicitly]
+			private set
+			{
+				if (value == m_PrivacyMuted)
+					return;
+
+				m_PrivacyMuted = value;
+
+				OnPrivacyMuteChanged.Raise(this, new BoolEventArgs(m_PrivacyMuted));
+			}
+		}
+
+		/// <summary>
+		/// Gets the DoNotDisturb state.
+		/// </summary>
+		public bool DoNotDisturb
+		{
+			get { return m_DoNotDisturb; }
+			[UsedImplicitly]
+			private set
+			{
+				if (value == m_DoNotDisturb)
+					return;
+
+				m_DoNotDisturb = value;
+
+				OnDoNotDisturbChanged.Raise(this, new BoolEventArgs(m_DoNotDisturb));
+			}
+		}
+
+		/// <summary>
+		/// Gets the type of conference this dialer supports.
+		/// </summary>
+		public eConferenceSourceType Supports { get; [UsedImplicitly] private set; }
+
+		#endregion
+
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		protected AbstractProxyDialingDeviceControl(IProxyDeviceBase parent, int id)
+			: base(parent, id)
+		{
+		}
+
+		/// <summary>
+		/// Override to release resources.
+		/// </summary>
+		/// <param name="disposing"></param>
+		protected override void DisposeFinal(bool disposing)
+		{
+			OnSourceAdded = null;
+			OnDoNotDisturbChanged = null;
+			OnAutoAnswerChanged = null;
+			OnPrivacyMuteChanged = null;
+
+			base.DisposeFinal(disposing);
+		}
+
+		#region Methods
+
+		/// <summary>
+		/// Gets the active conference sources.
+		/// </summary>
+		/// <returns></returns>
+		public IEnumerable<IConferenceSource> GetSources()
+		{
+			// TODO
+			yield break;
+		}
+
+		/// <summary>
+		/// Dials the given number.
+		/// </summary>
+		/// <param name="number"></param>
+		public void Dial(string number)
+		{
+			CallMethod(DialingDeviceControlApi.METHOD_DIAL);
+		}
+
+		/// <summary>
+		/// Dials the given number.
+		/// </summary>
+		/// <param name="number"></param>
+		/// <param name="callType"></param>
+		public void Dial(string number, eConferenceSourceType callType)
+		{
+			CallMethod(DialingDeviceControlApi.METHOD_DIAL_TYPE, callType);
+		}
+
+		/// <summary>
+		/// Sets the do-not-disturb enabled state.
+		/// </summary>
+		/// <param name="enabled"></param>
+		public void SetDoNotDisturb(bool enabled)
+		{
+			CallMethod(DialingDeviceControlApi.METHOD_SET_DO_NOT_DISTURB, enabled);
+		}
+
+		/// <summary>
+		/// Sets the auto-answer enabled state.
+		/// </summary>
+		/// <param name="enabled"></param>
+		public void SetAutoAnswer(bool enabled)
+		{
+			CallMethod(DialingDeviceControlApi.METHOD_SET_AUTO_ANSWER, enabled);
+		}
+
+		/// <summary>
+		/// Sets the privacy mute enabled state.
+		/// </summary>
+		/// <param name="enabled"></param>
+		public void SetPrivacyMute(bool enabled)
+		{
+			CallMethod(DialingDeviceControlApi.METHOD_SET_PRIVACY_MUTE, enabled);
+		}
+
+		#endregion
+	}
+}